modified code based on inputs - removed 'nullable' for specific fields, added 'Autowired' annotation for constructor, naming conventions for controller path, updated Account repository interface
