+++ src/main/java/com/amigoscode/group/ebankingsuite/account/AccountRepository.java
@@ -1,11 +1,11 @@
-package com.amigoscode.group.ebankingsuite.Account;
+package com.amigoscode.group.ebankingsuite.account;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Integer> {
-    List<Account> findByUserId(Integer userId);
+    Optional<Account> findByUserId(Integer userId);
 }
